Extract quadtree logic into separate module
