+++ check_register/__init__.py
@@ -1,9 +1,5 @@
 from .models import CheckEntry, RowChunk
 from .parser import CheckRegisterParser
-from .outputs import (
-    write_csv,
-    write_json,
-    write_payee_quadtree_html,
-    write_chunks,
-)
+from .outputs import write_csv, write_json, write_chunks
+from .quadtree import write_payee_quadtree_html
 from .stats import sanity, month_rollups, month_totals
